# Chainsaw: Partial rewrite, various improvements