Introduces protection checks for the entire tree
More efficient node digging (VoxelManip)
Improved drop handling using detached inventories for correct stack sizes


TODO:

 - [x] Test on the giant sequoia
 - [x] Performance difference measurements

PR giant sequoia:
```
chainsaw_dig: 54800 us
chainsaw_dig: 55004 us
chainsaw_dig: 49445 us


handle_drops: 4969 us
handle_drops: 3776 us
handle_drops: 2130 us
```

Master giant sequoia:
```
chainsaw_dig: 400130 us
chainsaw_dig: 410763 us
chainsaw_dig: 376324 us
```

Master takes approx. 7.3 times longer to execute.